1.Fixed on defect, should add some data before testing. 2. Formatted code. Welcome to test
git-svn-id: http://svn.code.sf.net/p/fossology/code/branches/fossology2.0/fossology@4614 a5643f74-2041-0410-a2b9-94ffef304a07

+++ src/wget_agent/agent_tests/Unit/wget_agent/testDBLoadGold.c
@@ -27,6 +27,8 @@ with this program; if not, write to the Free Software Foundation, Inc.,
  * \brief testing for the function DBLoadGold
  */
 
+static PGresult *result = NULL;
+
 /**
  * \brief initialize
  * at first download one file(dir), save as one tar file
@@ -43,13 +45,48 @@ int  DBLoadGoldInit()
   GetURL(TempFile, URL, TempFileDir);
   strcpy(GlobalTempFile,"./test_result/wget.tar");
   strcpy(GlobalURL, "http://fossology.org/debian/1.0.0/");
-  GlobalUploadKey = 1;
   pgConn = fo_dbconnect();
   if (!pgConn)
   {
     FATAL("Unable to connect to database");
     SafeExit(20);
   }
+  /** delete the record that upload_filename is wget.tar, pre testing */
+  memset(SQL,'\0',MAXCMD);
+  snprintf(SQL,MAXCMD, "DELETE FROM upload where upload_filename = 'wget.tar';");
+  result =  PQexec(pgConn, SQL);
+  if (fo_checkPQcommand(pgConn, result, SQL, __FILE__ ,__LINE__))
+  {
+    printf("Perpare delete information ERROR!\n");
+    if (pgConn) PQfinish(pgConn);
+    exit(-1);
+  }
+  PQclear(result);
+
+  /** insert upload wget.tar */
+  memset(SQL,'\0',MAXCMD);
+  snprintf(SQL,MAXCMD,"INSERT INTO upload (upload_filename,upload_mode,upload_ts) VALUES ('wget.tar',40,now());");
+  result =  PQexec(pgConn, SQL);
+  if (fo_checkPQcommand(pgConn, result, SQL, __FILE__ ,__LINE__))
+  {
+    printf("Perpare upload information ERROR!\n");
+    if (pgConn) PQfinish(pgConn);
+    exit(-1);
+  }
+  PQclear(result);
+  /** get upload id */
+  memset(SQL,'\0',MAXCMD);
+  snprintf(SQL,MAXCMD,"SELECT upload_pk from upload where upload_filename = 'wget.tar';");
+  result =  PQexec(pgConn, SQL);
+  if (fo_checkPQresult(pgConn, result, SQL, __FILE__ ,__LINE__))
+  {
+    printf("Perpare upload information ERROR!\n");
+    if (pgConn) PQfinish(pgConn);
+    exit(-1);
+  }
+  GlobalUploadKey = atoi(PQgetvalue(result,0,0));
+  PQclear(result);
+
   return 0;
 }
 /**
@@ -58,12 +95,24 @@ int  DBLoadGoldInit()
 int DBLoadGoldClean()
 {
   GlobalUploadKey = -1;
-	memset(GlobalTempFile, 0, MAXCMD);
-	memset(GlobalURL, 0, MAXCMD);
-	memset(GlobalParam, 0, MAXCMD);
+  memset(GlobalTempFile, 0, MAXCMD);
+  memset(GlobalURL, 0, MAXCMD);
+  memset(GlobalParam, 0, MAXCMD);
   char TempFileDir[MAXCMD];
-	if (pgConn) PQfinish(pgConn);
-	strcpy(TempFileDir, "./test_result");
+  /** delete the record that upload_filename is wget.tar, post testing */
+  memset(SQL,'\0',MAXCMD);
+  snprintf(SQL,MAXCMD, "DELETE FROM upload where upload_filename = 'wget.tar';");
+  result =  PQexec(pgConn, SQL);
+  if (fo_checkPQcommand(pgConn, result, SQL, __FILE__ ,__LINE__))
+  {
+    printf("Perpare delete information ERROR!\n");
+    if (pgConn) PQfinish(pgConn);
+    exit(-1);
+  }
+  PQclear(result);
+
+  if (pgConn) PQfinish(pgConn);
+  strcpy(TempFileDir, "./test_result");
   if (file_dir_existed(TempFileDir))
   {
     RemoveDir(TempFileDir);
@@ -78,8 +127,8 @@ int DBLoadGoldClean()
  */
 void string_tolower(char *string)
 {
-	int length = strlen(string);  
-	int i = 0;
+  int length = strlen(string);  
+  int i = 0;
   for (i = 0; i < length; i++)  
   {  
     string[i] = tolower(string[i]);  
@@ -93,54 +142,54 @@ void string_tolower(char *string)
  */
 void testDBLoadGold()
 {
-	printf("db start\n");
-	DBLoadGold();
-	printf("db end\n");
-	char SQL[MAXCMD];
-	char *pfile_sha1;
-	char *pfile_md5;
+  printf("db start\n");
+  DBLoadGold();
+  printf("db end\n");
+  char SQL[MAXCMD];
+  char *pfile_sha1;
+  char *pfile_md5;
   memset(SQL, 0, MAXCMD);
-	PGresult *result;
-	strcpy(SQL, "select pfile_sha1, pfile_md5 from pfile where pfile_pk in (select pfile_fk	from upload where upload_pk = 1);");
+  PGresult *result;
+  snprintf(SQL, MAXCMD-1, "select pfile_sha1, pfile_md5 from pfile where pfile_pk in (select pfile_fk from upload where upload_pk = %ld);", GlobalUploadKey);
   result =  PQexec(pgConn, SQL); /* SELECT */
   if (fo_checkPQresult(pgConn, result, SQL, __FILE__, __LINE__))
   {
     SafeExit(-1);
   }
   pfile_sha1 = PQgetvalue(result,0,0);
-	pfile_md5 = PQgetvalue(result,0,1);
+  pfile_md5 = PQgetvalue(result,0,1);
   printf("pfile_sha1, pfile_md5 are:%s, %s\n", pfile_sha1, pfile_md5 );
   string_tolower(pfile_sha1);
   string_tolower(pfile_md5);
   printf("pfile_sha1, pfile_md5 are:%s, %s\n", pfile_sha1, pfile_md5 );
-	char file_name_file[MAXCMD] = {0};
-	char file_name_gold[MAXCMD] = {0};
-	char string0[3] = {0};
-	char string1[3] = {0};
-	char string2[3] = {0};
-	char string4[] = "/srv/fossology/repository/localhost";
+  char file_name_file[MAXCMD] = {0};
+  char file_name_gold[MAXCMD] = {0};
+  char string0[3] = {0};
+  char string1[3] = {0};
+  char string2[3] = {0};
+  char string4[] = "/srv/fossology/repository/localhost";
   strncpy(string0, pfile_sha1, 2);
   strncpy(string1, pfile_sha1 + 2, 2);
   strncpy(string2, pfile_sha1 + 4, 2);
-	printf("string0, string1, string2 are:%s, %s, %s\n", string0, string1, string2);
-	sprintf(file_name_file, "%s/files/%s/%s/%s/%s.%s.10240", string4, string0, string1, string2, pfile_sha1, pfile_md5);
-	sprintf(file_name_gold, "%s/gold/%s/%s/%s/%s.%s.10240", string4, string0, string1, string2, pfile_sha1, pfile_md5);
-	int existed = file_dir_existed(file_name_file);
-	CU_ASSERT_EQUAL(existed, 1); /* the file into repo? */
+  printf("string0, string1, string2 are:%s, %s, %s\n", string0, string1, string2);
+  sprintf(file_name_file, "%s/files/%s/%s/%s/%s.%s.10240", string4, string0, string1, string2, pfile_sha1, pfile_md5);
+  sprintf(file_name_gold, "%s/gold/%s/%s/%s/%s.%s.10240", string4, string0, string1, string2, pfile_sha1, pfile_md5);
+  int existed = file_dir_existed(file_name_file);
+  CU_ASSERT_EQUAL(existed, 1); /* the file into repo? */
   if (existed)
   {
     RemoveDir(file_name_file);
   }
-	existed = 0;
-	existed = file_dir_existed(file_name_gold);
-	CU_ASSERT_EQUAL(existed, 1); /* the file into repo? */
-	printf("file_name_file, file_name_gold are:%s,%s\n", file_name_file, file_name_gold);
+  existed = 0;
+  existed = file_dir_existed(file_name_gold);
+  CU_ASSERT_EQUAL(existed, 1); /* the file into repo? */
+  printf("file_name_file, file_name_gold are:%s,%s\n", file_name_file, file_name_gold);
   if (existed)
   {
     RemoveDir(file_name_gold);
   }
-	PQclear(result);
-	printf("testDBLoadGold end\n");
+  PQclear(result);
+  printf("testDBLoadGold end\n");
 }
 
 /**
